@@ -13,8 +13,9 @@ import platform as platform_lib
 import socket
 import struct
 import subprocess as sp
-
 import warnings
+
+import winreg
 
 from scapy.arch.windows.structures import _windows_title, \
     GetAdaptersAddresses, GetIpForwardTable, GetIpForwardTable2, \
@@ -34,8 +35,6 @@ from scapy.pton_ntop import inet_ntop, inet_pton
 from scapy.utils import atol, itom, mac2str, str2mac
 from scapy.utils6 import construct_source_candidate_set, in6_getscope
 from scapy.data import ARPHDR_ETHER, load_manuf
-import scapy.libs.six as six
-from scapy.libs.six.moves import input, winreg
 from scapy.compat import plain_str
 from scapy.supersocket import SuperSocket
 
@@ -298,16 +297,12 @@ def get_windows_if_list(extended=False):
             ips.extend(_resolve_ips(multicast))
         return ips
 
-    if six.PY2:
-        _str_decode = lambda x: x.encode('utf8', errors='ignore')
-    else:
-        _str_decode = plain_str
     return [
         {
-            "name": _str_decode(x["friendly_name"]),
+            "name": plain_str(x["friendly_name"]),
             "index": x["interface_index"],
-            "description": _str_decode(x["description"]),
-            "guid": _str_decode(x["adapter_name"]),
+            "description": plain_str(x["description"]),
+            "guid": plain_str(x["adapter_name"]),
             "mac": _get_mac(x),
             "ipv4_metric": 0 if WINDOWS_XP else x["ipv4_metric"],
             "ipv6_metric": 0 if WINDOWS_XP else x["ipv6_metric"],
@@ -623,7 +618,7 @@ class WindowsInterfacesProvider(InterfaceProvider):
                 windows_interfaces[i['guid']] = i
 
         index = 0
-        for netw, if_data in six.iteritems(conf.cache_pcapiflist):
+        for netw, if_data in conf.cache_pcapiflist.items():
             name, ips, flags, _ = if_data
             guid = _pcapname_to_guid(netw)
             data = windows_interfaces.get(guid, None)
@@ -674,7 +669,7 @@ def get_ips(v6=False):
     :param v6: IPv6 addresses
     """
     res = {}
-    for iface in six.itervalues(conf.ifaces):
+    for iface in conf.ifaces.values():
         if v6:
             res[iface] = iface.ips[6]
         else:
