@@ -8,8 +8,6 @@
 ASN.1 (Abstract Syntax Notation One)
 """
 
-from __future__ import absolute_import
-from __future__ import print_function
 import random
 
 from datetime import datetime, timedelta, tzinfo
@@ -18,7 +16,6 @@ from scapy.error import Scapy_Exception, warning
 from scapy.volatile import RandField, RandIP, GeneralizedTime
 from scapy.utils import Enum_metaclass, EnumElement, binrepr
 from scapy.compat import plain_str, bytes_encode, chb, orb
-import scapy.libs.six as six
 
 from scapy.compat import (
     Any,
@@ -79,9 +76,7 @@ class RandASN1Object(RandField["ASN1_Object[Any]"]):
         else:
             self.objlist = [
                 x._asn1_obj
-                for x in six.itervalues(
-                    ASN1_Class_UNIVERSAL.__rdict__  # type: ignore
-                )
+                for x in ASN1_Class_UNIVERSAL.__rdict__.values()  # type: ignore
                 if hasattr(x, "_asn1_obj")
             ]
         self.chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789"  # noqa: E501
@@ -149,8 +144,7 @@ class ASN1_Codecs_metaclass(Enum_metaclass):
     element_class = ASN1Codec
 
 
-@six.add_metaclass(ASN1_Codecs_metaclass)
-class ASN1_Codecs:
+class ASN1_Codecs(metaclass=ASN1_Codecs_metaclass):
     BER = cast(ASN1Codec, 1)
     DER = cast(ASN1Codec, 2)
     PER = cast(ASN1Codec, 3)
@@ -215,12 +209,12 @@ class ASN1_Class_metaclass(Enum_metaclass):
                 ):
         # type: (...) -> Type[ASN1_Class]
         for b in bases:
-            for k, v in six.iteritems(b.__dict__):
+            for k, v in b.__dict__.items():
                 if k not in dct and isinstance(v, ASN1Tag):
                     dct[k] = v.clone()
 
         rdict = {}
-        for k, v in six.iteritems(dct):
+        for k, v in dct.items():
             if isinstance(v, int):
                 v = ASN1Tag(k, v)
                 dct[k] = v
@@ -231,15 +225,14 @@ class ASN1_Class_metaclass(Enum_metaclass):
 
         ncls = cast('Type[ASN1_Class]',
                     type.__new__(cls, name, bases, dct))
-        for v in six.itervalues(ncls.__dict__):
+        for v in ncls.__dict__.values():
             if isinstance(v, ASN1Tag):
                 # overwrite ASN1Tag contexts, even cloned ones
                 v.context = ncls
         return ncls
 
 
-@six.add_metaclass(ASN1_Class_metaclass)
-class ASN1_Class:
+class ASN1_Class(metaclass=ASN1_Class_metaclass):
     pass
 
 
@@ -306,8 +299,7 @@ class ASN1_Object_metaclass(_Generic_metaclass):
 _K = TypeVar('_K')
 
 
-@six.add_metaclass(ASN1_Object_metaclass)
-class ASN1_Object(Generic[_K]):
+class ASN1_Object(Generic[_K], metaclass=ASN1_Object_metaclass):
     tag = ASN1_Class_UNIVERSAL.ANY
 
     def __init__(self, val):
