@@ -14,7 +14,6 @@ Specifications can be retrieved from https://www.opennetworking.org/
 # scapy.contrib.description = Openflow v1.0
 # scapy.contrib.status = loads
 
-from __future__ import absolute_import
 import struct
 
 
@@ -26,7 +25,6 @@ from scapy.layers.l2 import Ether
 from scapy.layers.inet import TCP
 from scapy.packet import Packet, Raw, bind_bottom_up, bind_top_down
 from scapy.utils import binrepr
-import scapy.libs.six as six
 
 
 # If prereq_autocomplete is True then match prerequisites will be
@@ -717,7 +715,7 @@ class OFPTFeaturesRequest(_ofp_header):
                    IntField("xid", 0)]
 
 
-ofp_action_types_flags = [v for v in six.itervalues(ofp_action_types)
+ofp_action_types_flags = [v for v in ofp_action_types.values()
                           if v != 'OFPAT_VENDOR']
 
 
