@@ -17,7 +17,6 @@ from scapy.compat import Optional, List, Type, Any, Tuple, Iterable, Dict, \
     cast, Callable, orb
 from scapy.contrib.automotive import log_automotive
 from scapy.packet import Packet
-import scapy.libs.six as six
 from scapy.config import conf
 from scapy.supersocket import SuperSocket
 from scapy.error import Scapy_Exception
@@ -52,8 +51,7 @@ __all__ = ["GMLAN_Scanner", "GMLAN_ServiceEnumerator", "GMLAN_RDBIEnumerator",
            "GMLAN_DCEnumerator"]
 
 
-@six.add_metaclass(abc.ABCMeta)
-class GMLAN_Enumerator(ServiceEnumerator):
+class GMLAN_Enumerator(ServiceEnumerator, metaclass=abc.ABCMeta):
     """
     Abstract base class for GMLAN service enumerators. This class
     implements GMLAN specific functions.
