@@ -8,10 +8,9 @@ PacketList: holds several packets and allows to do operations on them.
 """
 
 
-from __future__ import absolute_import
-from __future__ import print_function
 import os
 from collections import defaultdict
+from typing import NamedTuple
 
 from scapy.compat import lambda_tuple_converter
 from scapy.config import conf
@@ -36,7 +35,6 @@ from scapy.compat import (
     Generic,
     Iterator,
     List,
-    NamedTuple,
     Optional,
     Tuple,
     Type,
